// src/pages/Users/type.ts
export interface UserDataType {
  username: string;
  email: string;
  phone_number: string;
  organization: string;
  first_name: string;
  last_name: string;
  date_joined: string;
  gender: string;
  BVN: string;
  facebook: string;
  instagram: string;
  twitter: string;
  guarantor_name: string;
  guarantor_phone_number: string;
  guarantor_email: string;
  status: string;
  amount: number;
  acount: string;
  bank: string;
  star: number;
  tag: string;
}